+++ test/metadata.test.js
@@ -12,7 +12,7 @@ describe('parquetMetadata', () => {
     it(`parse metadata from ${file}`, async () => {
       const asyncBuffer = await asyncBufferFromFile(`test/files/${file}`)
       const arrayBuffer = await asyncBuffer.slice(0)
-      const result = toJson(parquetMetadataSync(arrayBuffer))
+      const result = toJson(parquetMetadataSync({ file: arrayBuffer }))
       const base = file.replace('.parquet', '')
       const expected = fileToJson(`test/files/${base}.metadata.json`)
       expect(result, JSON.stringify(result, null, 2)).toEqual(expected)
@@ -20,34 +20,35 @@ describe('parquetMetadata', () => {
   })
 
   it('throws for arrayBuffer undefined', () => {
+    const file = undefined
     // @ts-expect-error testing invalid input
-    expect(() => parquetMetadataSync(undefined)).toThrow('parquet expected ArrayBuffer')
+    expect(() => parquetMetadataSync({ file })).toThrow('parquet expected ArrayBuffer')
   })
 
   it('throws for a too short file', () => {
-    const arrayBuffer = new ArrayBuffer(0)
-    expect(() => parquetMetadataSync(arrayBuffer)).toThrow('parquet file is too short')
+    const file = new ArrayBuffer(0)
+    expect(() => parquetMetadataSync({ file })).toThrow('parquet file is too short')
   })
 
   it('throws for invalid metadata length', () => {
-    const arrayBuffer = new ArrayBuffer(12)
-    const view = new DataView(arrayBuffer)
+    const file = new ArrayBuffer(12)
+    const view = new DataView(file)
     view.setUint32(0, 0x31524150, true) // magic number PAR1
     view.setUint32(4, 1000, true) // 1000 bytes exceeds buffer
     view.setUint32(8, 0x31524150, true) // magic number PAR1
-    expect(() => parquetMetadataSync(arrayBuffer))
+    expect(() => parquetMetadataSync({ file }))
       .toThrow('parquet metadata length 1000 exceeds available buffer 4')
   })
 
   it('throws for invalid magic number', () => {
-    const arrayBuffer = new ArrayBuffer(8)
-    expect(() => parquetMetadataSync(arrayBuffer))
+    const file = new ArrayBuffer(8)
+    expect(() => parquetMetadataSync({ file }))
       .toThrow('parquet file invalid (footer != PAR1)')
   })
 
-  it('throws for invalid metadata length', () => {
+  it('throws for invalid metadata length 2', () => {
     const { buffer } = new Uint8Array([255, 255, 255, 255, 80, 65, 82, 49])
-    expect(() => parquetMetadataSync(buffer))
+    expect(() => parquetMetadataSync({ file: buffer }))
       .toThrow('parquet metadata length 4294967295 exceeds available buffer 0')
   })
 })
@@ -56,7 +57,7 @@ describe('parquetMetadataAsync', () => {
   files.forEach(file => {
     it(`parse metadata async from ${file}`, async () => {
       const asyncBuffer = await asyncBufferFromFile(`test/files/${file}`)
-      const result = await parquetMetadata(asyncBuffer)
+      const result = await parquetMetadata({ file: asyncBuffer })
       const base = file.replace('.parquet', '')
       const expected = fileToJson(`test/files/${base}.metadata.json`)
       expect(toJson(result)).toEqual(expected)
@@ -64,21 +65,21 @@ describe('parquetMetadataAsync', () => {
   })
 
   it('throws for asyncBuffer undefined', async () => {
-    const arrayBuffer = undefined
+    const file = undefined
     // @ts-expect-error testing invalid input
-    await expect(parquetMetadata(arrayBuffer)).rejects
+    await expect(parquetMetadata({ file })).rejects
       .toThrow('parquet expected AsyncBuffer')
   })
 
   it('throws for invalid magic number', async () => {
     const { buffer } = new Uint8Array([255, 255, 255, 255, 255, 255, 255, 255])
-    await expect(parquetMetadata(buffer)).rejects
+    await expect(parquetMetadata({ file: buffer })).rejects
       .toThrow('parquet file invalid (footer != PAR1)')
   })
 
   it('throws for invalid metadata length', async () => {
     const { buffer } = new Uint8Array([255, 255, 255, 255, 80, 65, 82, 49])
-    await expect(parquetMetadata(buffer)).rejects
+    await expect(parquetMetadata({ file: buffer })).rejects
       .toThrow('parquet metadata length 4294967295 exceeds available buffer 0')
   })
 })