+++ test/readFiles.test.js
@@ -35,7 +35,7 @@ describe('parquetRead test files', () => {
     it(`read the last row from ${filename}`, async () => {
       // this exercises some of the page-skipping optimizations
       const file = await asyncBufferFromFile(`test/files/${filename}`)
-      const metadata = await parquetMetadata(file)
+      const metadata = await parquetMetadata({ file })
       let numRows = Number(metadata.num_rows)
       // repeated_no_annotation has wrong num_rows in metadata:
       if (filename === 'repeated_no_annotation.parquet') numRows = 6